+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ```
     * In AbstractRequest class, getParams() method, was added additional check for some of the merchant details.
     * In ConfigProvider class, will typecast the "sdk_style" to string.
+    * Fixed the problem in OpenOrder class, where the apmWindowType value was wrong. This prevented call of updateOrder.
 ```
 
 # 3.3.1